feat: Add 'ip' subcommand to print resolved hostname
This commit introduces a new 'ip' subcommand to the `remote` utility.
When executed as `remote ip`, it will resolve the remote hostname
(using the existing configuration and caching mechanism) and print it
to standard output. This allows users to easily retrieve the IP address
or hostname that `remote` would use for connections without initiating
an SSH or Rsync operation.