refactor: Use runtime/debug to get version info
This commit refactors the way the `remote` command retrieves its
version information. Previously, the version was injected at build
time using ldflags in the Makefile. This approach, while functional,
is not compatible with `go install`.

This change updates the code to use `runtime/debug.ReadBuildInfo()`
to get version information. This allows the binary to be installed
with `go install` and still report the correct version.

The Makefile has been updated to remove the ldflags, and the code
has been updated to use the new method.

Additionally, this commit includes fixes for various issues
reported by `golang-ci lint`, including unchecked errors and
improving error formatting.